+++ safe-coloured-text-parsing/src/Text/Colour/Chunk/Parsing.hs
@@ -236,6 +236,8 @@ tokensToChunks style tokens =
              in (finalS, Chunk {chunkText = t, chunkStyle = s} : restChunks)
       SgrSequence params -> go (applySGRParams s params) rest
       OtherCsiSequence -> go s rest
+      OscSequence (OscHyperlink _ url) ->
+        go (s {chunkStyleHyperlink = Just url}) rest
       OscSequence _ -> go s rest
       CarriageReturn ->
         case rest of